+++ runtime/core/xnix_runtime_core_cli.c
@@ -2829,6 +2829,8 @@ print_probe(void)
   fputs("\"compatibility_test_result_type_count\":3,", stdout);
   fputs("\"execution_readiness_owner\":\"c\",", stdout);
   fputs("\"execution_readiness_application_count\":1,", stdout);
+  fputs("\"launch_intent_owner\":\"c\",", stdout);
+  fputs("\"launch_intent_application_count\":1,", stdout);
   fputs("\"desktop_activation_manifest_owner\":\"c\",", stdout);
   fputs("\"desktop_activation_manifest_count\":1,", stdout);
   fputs("\"kde_integration_status_owner\":\"c\",", stdout);
@@ -2923,6 +2925,7 @@ print_write_gate(const char *method_name)
 #include "xnix_runtime_core_cli_compatibility_test_plan.inc"
 #include "xnix_runtime_core_cli_compatibility_test_result.inc"
 #include "xnix_runtime_core_cli_execution_readiness.inc"
+#include "xnix_runtime_core_cli_launch_intent.inc"
 #include "xnix_runtime_core_cli_compatibility_repair_plan.inc"
 #include "xnix_runtime_core_cli_ai_diagnostic_input.inc"
 #include "xnix_runtime_core_cli_ai_diagnostic_recommendation.inc"