+++ test/test_runtime_contract.rb
@@ -45,6 +45,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetTestPlan
   GetTestResult
   GetExecutionReadiness
+  GetLaunchIntent
   GetAIDiagnosticInput
   GetAIDiagnosticRecommendation
   GetAIRepairApprovalGate